backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pyserial==3.5
python-multipart==0.0.9
websockets==12.0
